Fix url for libturbo and update info (#578)

+++ scripts/start.sh
@@ -152,7 +152,9 @@ export  READONLY CLUSTER_DURABILITY BROKER_URL EMBEDDED_BROKER TOGGLE_MASTER TOG
 export CLUSTER_CONFIG_DIR MONITOR_AUDIT_PATH INSTANCE_STRING
 # Cleanup existing monitoring files
 if [[ ${CLUSTERING} =~ [Ff][Aa][Ll][Ss][Ee] ]]; then
-find "${GEOSERVER_DATA_DIR}"/monitoring -type d -name 'monitor_*' -exec rm -r {} +
+  if [[ -d "${GEOSERVER_DATA_DIR}"/monitoring ]];then
+    find "${GEOSERVER_DATA_DIR}"/monitoring -type d -name 'monitor_*' -exec rm -r {} +
+  fi
 fi
 create_dir "${MONITOR_AUDIT_PATH}"
 setup_monitoring
